audit: tweak assert/refute_predicate audit
Addresses some of the issues I raised in https://github.com/Homebrew/homebrew-core/pull/19127#issue-263566817.
Would still like to work out a way to check `File.exist?`-style language
inside the `test do` block exclusively but for now it's better to be
too conservative on the audit than to be overly-zealous and flag too
much.