formula: read runtime dependencies from more tabs.
The fixes the issue mentioned in:
https://github.com/Homebrew/brew/pull/4066#issuecomment-381308153
where `brew missing` was not reading from a tab of an outdated formula.

The `installed_prefix.directory?` check is unnecessary as the following
`opt_or_installed_prefix_keg` call is already checking that when it is
necessary. In this case it's not necessary anyway as `opt_prefix` is
preferred and makes more sense on outdated formulae.