Increase spec timeout to 30 seconds
I’ve noticed the deadlock test in `cask/system_command_spec.rb` appears to fail spuriously, e. g. in #4529.

Details:

https://travis-ci.org/Homebrew/brew/jobs/406827725#L297-L318

With the spec taking 2 seconds on average on my 2017 MBP, I feel it’s
entirely plausible for the aging test-bot hardware to get knocked by
the 15-second timeout.

Given that the spec will pass early anyway if it passes, I feel it’s
reasonable to bump the timeout to 30 seconds.