bump-formula-pr: add --no-fork
GitHub seems to be discouraging forking private repositories[0]:

> By default, new organizations are configured to disallow the forking
> of private repositories.

bump-formula-pr tries to create its pull requests from a fork, so it
can't be used for private taps set up in this way.

I've added a --no-fork option that will create PRs in the tap repo
itself, rather than in a fork, to accommodate this use case.

[0]: https://help.github.com/articles/allowing-people-to-fork-private-repositories-in-your-organization/