Swap in description of current behavior instead of 1.9.3 hack comment
While working on 40c71c122, I noticed this odd comment and decided to
investigate. I wasn't able to reproduce the undesirable behavior because
I cannot get 1.8.7 to compile with rvm but I did notice through some
testing that iTerm2 displays a lock character when echo mode is
disabled. My intent had been to remove this method because Homebrew has
moved past a Ruby version that would necessitate the "hack" but this
method's behavior is actually desirable today, so I changed the comment
to remove the potential flag and instead left a comment explaining
a good second reason for the method.