cask/config_spec: add failing test
We recently added a new global artifact and then updated a cask to
make use of that new artifact. This caused a number of `brew cask`
commands to fail for users who had the cask installed before the
artifact was added.

This commit adds test cases for that failure mode.

See also:
- https://github.com/Homebrew/brew/pull/7286#issuecomment-613376568
- https://discourse.brew.sh/t/cask-definition-is-invalid-invalid-mdimporter-stanza-key-not-found-mdimporterdir