workflows/tests: consolidate macOS tests.
We have a limited number of macOS workers (for the Homebrew
organisation) and queuing these seems to be slowing things down overall
compared to when we had one big job.

Linux throughput is still significantly quicker so it makes sense to
split up those jobs for providing quick feedback on breakages that are
cross-platform (which ends up being most but definitely not all of
them).