dev-cmd/release: output blog post release notes.
I forgot until I started to work on the 3.0.0 release that there was an
added step I would do for a major/minor release:

```
brew release-notes --markdown 2.7.0 | pbcopy
```

This would generate for 3.0.0/2.8.0 the release notes since 2.7.0 for
editing down into a blog post.

Add this behaviour to `brew release` and, because we're doing it
programmatically anyway, take this chance to also reject Dependabot
lines and sort them for easier grouping by pull request subject.