Stop including `*flight` block source in cask API
Originally we were going to try and load the *flight blocks from the API
but we ended up going with downloading the caskfile for the subset of
casks that need this functionality for consisty's sake.

This reverts the following commits from most recent to oldest:
- 36dbbd8fd1f7c9e249bbbbd12004edf52eccaef4
- 6d6c8eaed0e7980043c5ab9eed5a750dc9769fe8
- bdb0cb96885e6cc0ead1f957d8c1be6b6c609308
- 228b87a4ecf7fc0d51db9f1a809934e96dd0b743
- c2fe3867f470e41d66379c0efff04e635a092852

It also changes how *flight blocks are handled in `.to_h`.
Essentially, when *flight blocks exist they are just included
as a hash of the artifact to nil to indicate that they exist.

More information isn't necessary since we don't evaluate
the current source code in the *flight artifacts that
we get from the API.